requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
autopep8==2.0.1
billiard==3.6.4.0
bravado-core==5.17.1
bravado==10.6.3
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
crispy-bootstrap5==0.7
django-bitfield==2.2.0
django-bootstrap-datepicker-plus==5.0.3
django-celery-beat==2.4.0
django-colorfield==0.8.0
django-crispy-forms==2.0
django-esi==4.0.1
django-eve-auth==1.2.0
django-eveuniverse==0.19.0
django-extensions==3.2.1
django-timezone-field==5.0
django==4.1.5
ecdsa==0.18.0
frozenlist==1.3.3
gitdb==4.0.10
gitpython==3.1.31
gunicorn==20.1.0
idna==3.4
jsonpointer==2.3
jsonref==1.1.0
jsonschema==3.2.0
kombu==5.2.4
monotonic==1.6
msgpack==1.0.4
multidict==6.0.4
oauthlib==3.2.2
pillow==10.0.0; python_version >= '3.8'
prompt-toolkit==3.0.36
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.4
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-discord-client==0.7.3.8
python-jose==3.3.0
pytz==2022.7.1
pyyaml==6.0
redis==4.4.2
requests-oauthlib==1.3.1
requests==2.28.2
rfc3987==1.3.8
rsa==4.9
setuptools==68.0.0; python_version >= '3.7'
simplejson==3.18.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
strict-rfc3339==0.7
swagger-spec-validator==3.0.3
tomli==2.0.1
tqdm==4.64.1
typing-extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webcolors==1.12
websockets==10.4
yarl==1.8.2
discord-webhook==1.1.0